requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.119.0
fastuuid==0.13.5
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
transformers==4.44.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.78.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.17.0
mdurl==0.1.2
multidict==6.7.0
networkx==3.5
numpy==2.2.6
openai==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
torch==2.3.1
timm==1.0.11
propcache==0.4.1
psutil==7.1.0
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pypdf==5.1.0
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
schedule==1.2.2
setuptools==80.9.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
swarms==8.5.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
win32_setctime==1.2.0
winloop==0.3.0
yarl==1.22.0
zipp==3.23.0